pyproject.toml
[tool.poetry]
name = "pseudotv"
version = "0.1.0"
description = ""
authors = ["jack"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
playwright = "^1.52.0"
python-dotenv = "^1.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
